Refactor: 기능명세서대로 폴더 및 파일 명 구성
