+++ src/index.js
@@ -1,4 +1,4 @@
-import App from "./App.js";
+import App from './controller/App.js';
 
 const app = new App();
 app.play();
